// src/app/shared/generated/model/scaffoldPageResult.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Scaffold } from './scaffold';


export interface ScaffoldPageResult { 
    totalElements?: number;
    number?: number;
    size?: number;
    totalPages?: number;
    stream?: Array<Scaffold>;
}