// src/app/shared/generated/model/skillPageResult.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Skill } from './skill';


export interface SkillPageResult { 
    totalElements?: number;
    number?: number;
    size?: number;
    totalPages?: number;
    stream?: Array<Skill>;
}